+++ server/routerlicious/packages/services-client/src/index.ts
@@ -4,7 +4,7 @@
  */
 
 export * from "./auth";
-export * from "./dockerNames";
+export * from "./generateNames";
 export * from "./gitManager";
 export * from "./historian";
 export * from "./interfaces";
